requirements.txt
Flask==3.0.0
requests==2.31.0
pywebview==4.4.1
gunicorn==21.2.0
